Got the game back in atleast somewhat good position :O
- Redid how game and clients/servers are setup. All of server, client, client char info
is done before the scene change. SteamLobby matchmaker does not tranfer any client
to server info. The server will wait for all players to load and to connect before
starting the game. (This all works with 1 player havent tested with multiple)

- IDs have been changed a bit. SteamIds are still in but are not used in any
essential systems. Most have been changed to ClientId if possible and then Player.id
which can be set to either steam/client ids but defaults to clientid. NetworkObjectId
is still used in cased where you want to get a specific object instead of a player.